use of event rendering buffer
